refactor: provide a server factory as a lib
